qspi: three firmware fixes
Three independent pre-existing bugs in the SPI / UART / parser
paths used by the qspi demo:

- common/spi.h:  uint16_t clkdiv in struct spi_cfg traps the
  SHARC+ core on the first halfword store into the stack-
  allocated config struct (FAULT LED latches).  cc21k emits
  a halfword store that the CPU refuses in the L1 block0
  stack mapping.  Promote clkdiv to uint32_t; callers already
  hand it a 32-bit value.

- common/uart.c: uart_try_getc never cleared the UART0_STAT
  error-latch bits (OE/PE/FE/BI).  Once any of them latches
  -- and OE latches routinely when the DSP is mid-printf and
  a new byte arrives -- the hardware stops asserting DR for
  subsequent bytes and RX dies silently.  Now W1C those bits,
  but only when DR was zero, because writing the status
  register while DR is set also clears DR as a side effect
  and we would eat the byte we just noticed.

- qspi/main.c: parse_seed_count parsed the seed with a
  decimal default base, so bare "c0ffee" (as the header
  comment advertises) was silently rejected and count ended
  up stuck at 0.  Add parse_u32_base with an explicit
  default-base argument, expose parse_u32 and parse_u32_hex
  wrappers, and use the hex variant for the seed.  0x-prefix
  override still works.

Also add qspi/TODO.md with the bring-up plan.  Step 0 in that
file lists these three fixes as prerequisites; Step 0.5
introduces a bring-up SPY loop (added in a follow-up commit)
for the SPI slave path.

+++ common/spi.h
@@ -31,7 +31,7 @@ enum spi_miom {
 
 // SPI port configuration.
 struct spi_cfg {
-   uint16_t clkdiv;         // baud = SCLK0 / (clkdiv + 1)
+   uint32_t clkdiv;         // baud = SCLK0 / (clkdiv + 1)
    enum spi_word_size size; // 8, 16, or 32-bit words
    enum spi_miom miom;      // single / dual / quad
    unsigned is_master : 1;  // 1 = master, 0 = slave
